+++ docs/changelog.rst
@@ -3,6 +3,7 @@ Changelog
 =========
 
 
+* :feature:`6733` Added support for detection of GRT tokens delegated to indexers in The Graph protocol (amounts including rewards).
 * :feature:`-` Binance CSV importing will now recognize more entry types.
 * :feature:`6712` The Graph protocol support has been added. The events related to delegator staking now will be properly displayed and accounted for.
 * :feature:`5843` Velodrome is now supported in Optimism. Related transactions should be decoded properly and shown in human readable format and any balances in Velodrome auto-detected.